serializer_drf_project/requirements.txt
asgiref==3.4.1
automium==0.2.6
automium-web==0.1.1
boto3==1.24.4
botocore==1.27.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.4
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==3.2
django-allauth==0.48.0
django-cors-headers==3.13.0
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filelock==3.4.2
Flask==2.0.3
gunicorn==20.1.0
httpie==3.2.1
idna==3.3
inflection==0.5.1
itsdangerous==2.1.0
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.0
legacy==0.1.6
Markdown==3.3.6
MarkupSafe==2.1.0
module==0.0.4
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.0
platformdirs==2.4.1
psycopg2==2.9.3
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.4.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
signals==0.0.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
userapp==1.0.5
views==0.3
virtualenv==20.13.0
virtualenvwrapper-win==1.2.7
waitress==2.1.2
Werkzeug==2.0.3
whitenoise==6.2.0
